student_management/app/config.py
import os

from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()


class Config:
    TESTING = True
    SQLALCHEMY_DATABASE_URI = os.getenv("SQLALCHEMY_DATABASE_URI")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = os.getenv('SECRET_KEY')
