requirements.txt
aniso8601==10.0.0
blinker==1.9.0
build==1.0.3
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
dulwich==0.21.6
Faker==35.0.0
fastjsonschema==2.18.1
filelock==3.16.1
Flask==3.1.0
Flask-RESTful==0.3.10
greenlet==3.1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==24.2.0
MarkupSafe==3.0.2
more-itertools==10.1.0
msgpack==1.0.7
packaging==24.1
pexpect==4.9.0
pipenv==2024.4.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
six==1.17.0
SQLAlchemy==2.0.37
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.12.2
urllib3==2.0.6
virtualenv==20.28.0
Werkzeug==3.1.3
zipp==3.17.0
